refactor: update import paths to use absolute paths and enhance build scripts with tsc-alias and tsconfig-paths
