+++ src/domain/use-cases/url/create-url.use-case.ts
@@ -1,4 +1,4 @@
-import { CreateUrlDto, UrlRepository } from '../..';
+import { CreateUrlDto, UrlRepository } from '@/domain';
 interface Url {
   id: string,
   name?: string,
