+++ src/infrastructure/mappers/user-google.mapper.ts
@@ -1,5 +1,5 @@
-import { CustomError, User } from '../../domain';
-import { Messages } from '../../config';
+import { CustomError, User } from '@/domain';
+import { Messages } from '@/config';
 
 export class UserGoogleMapper {
