+++ src/infrastructure/repositories/auth.repository.impl.ts
@@ -1,4 +1,4 @@
-import { AuthDataSource, AuthRepository, LoginUserDto, RegisterUserDto, ResetPasswordToken, UpdateUserDto, User } from '../../domain';
+import { AuthDataSource, AuthRepository, LoginUserDto, RegisterUserDto, ResetPasswordToken, UpdateUserDto, User } from '@/domain';
 
 export class AuthRepositoryImpl implements AuthRepository {
   // dependency injection 💉
