+++ src/infrastructure/repositories/url.repository.impl.ts
@@ -1,4 +1,4 @@
-import { CreateUrlDto, Url, UrlRepository, UrlDataSource, UpdateUrlDto, Page } from '../../domain';
+import { CreateUrlDto, Url, UrlRepository, UrlDataSource, UpdateUrlDto, Page } from '@/domain';
 
 export class UrlRepositoryImpl implements UrlRepository {
