+++ src/presentation/shortener/controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
-import { CustomError, UrlRepository } from '../../domain';
-import { RedirectUrl } from '../../domain';
-import { envs } from '../../config';
+import { CustomError, UrlRepository } from '@/domain';
+import { RedirectUrl } from '@/domain';
+import { envs } from '@/config';
 
 export class ShortenerController {
